Relax requirement to pass Table type
Permit use of instantiated Tables and raw table id strings in place
of Table types when creating query objects by removing the forced
auto-instantiation in v2/session.py's action-specific methods.

Instead, set the default repr of Table to be the table's ID string
(already done) and use that default repr rather than calling <table>.id.

This lets us use strs as well as Tables. Check that we're not attempting
this with a <type 'type'>; if so, do the instantiation as before, assuming
we've got a Table type.